requirements.txt
asgiref==3.8.1
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
inflection==0.5.1
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.3
uritemplate==4.1.1
